migrate provisioning module
