+++ src-tauri/enterprise/provisioning/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "defguard-provisioning"
+description = "Zero-touch provisioning for the Defguard desktop client"
+authors.workspace = true
+edition.workspace = true
+homepage.workspace = true
+license-file = "../LICENSE.md"
+rust-version.workspace = true
+version.workspace = true
+
+[dependencies]
+defguard-client-core = { path = "../../core" }
+log = { version = "0.4", features = ["serde"] }
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
