+++ src-tauri/src/enterprise/mod.rs
@@ -1,5 +1,6 @@
 pub use defguard_posture::inspector;
 pub use defguard_posture::posture;
+pub use defguard_provisioning::{try_get_provisioning_config, ProvisioningConfig};
 pub mod models;
 pub mod periodic;
 pub mod provisioning;
